+++ spec/models/user_spec.rb
@@ -242,6 +242,27 @@ RSpec.describe User, type: :model do
       it ".bottom_merchants_by_fulfillment_time" do
         expect(User.bottom_merchants_by_fulfillment_time(3)).to eq([@m2, @m3, @m6])
       end
+
+      it ".top_address_states_by_order_count" do
+        expect(Address.top_address_states_by_order_count(3)[0].state).to eq("IA")
+        expect(Address.top_address_states_by_order_count(3)[0].order_count).to eq(6)
+        expect(Address.top_address_states_by_order_count(3)[1].state).to eq("CO")
+        expect(Address.top_address_states_by_order_count(3)[1].order_count).to eq(3)
+        expect(Address.top_address_states_by_order_count(3)[2].state).to eq("OK")
+        expect(Address.top_address_states_by_order_count(3)[2].order_count).to eq(2)
+      end
+
+      it ".top_address_cities_by_order_count" do
+        expect(Address.top_address_cities_by_order_count(3)[0].state).to eq("IA")
+        expect(Address.top_address_cities_by_order_count(3)[0].city).to eq("Anywhere")
+        expect(Address.top_address_cities_by_order_count(3)[0].order_count).to eq(4)
+        expect(Address.top_address_cities_by_order_count(3)[1].state).to eq("CO")
+        expect(Address.top_address_cities_by_order_count(3)[1].city).to eq("Anywhere")
+        expect(Address.top_address_cities_by_order_count(3)[1].order_count).to eq(3)
+        expect(Address.top_address_cities_by_order_count(3)[2].state).to eq("IA")
+        expect(Address.top_address_cities_by_order_count(3)[2].city).to eq("Des Moines")
+        expect(Address.top_address_cities_by_order_count(3)[2].order_count).to eq(2)
+      end
     end
   end
 end
